Version One
In this version I set up the program with a pre-made recipe so I could focus on getting the scale factor to work. The user inputs how many people the recipe should serve, then the program divides this by how many people it originally served to find the scale factor. This is what all the values in the amounts list will be multiplied by.