Version Two
In this version I focused on the next component of the code – altering the amounts of each ingredient in the recipe using the scale factor. I looped through the amounts variable and multiplied each value in the list by the scale factor. This new value is appended to a list of all the new amounts. Then the new recipe is outputted, with each ingredient and its relevant unit and amount on one line. At this point, n/a still shows up.