Version Three
In version three I started to replace the hard-coded recipe with user input. For this version I was focusing on the ingredients and their units. I trialled using a menu-based input function, however I came across some issues with this. I decided that this was not the best way to get input from the user.