Version Four
In this version I replaced the menu-based input with a different method. If the user does not want to input an ingredient they press quit, otherwise it carries on to the units for them. This ensures that the lists are the same lengths, and the units and ingredients pair up well.