Version Five
In version five, I added the amounts and the original servings to the recipe_input function. I decided to do this in a different version because I had to make sure that I was still able to perform calculations on these amounts (they had to be float and int). Now that I had gotten rid of the hard-coded recipe, I could comment out the print statement crediting the original recipe. Also, now that the entire recipe was user-inputted, there were no more errors with the list length in output.