Version Six
In this version I added error handling for the boundaries. This was only relevant for the scale factor. So, if the new amount of servings that the user entered creates a scale factor that is greater than three or smaller than one quarter, the program stops them and prompts them to enter a different amount. I put the new servings input down with the scale factor calculation, and put both in a while True loop. This loops infinitely until the user enters a new serving amount that leads to a scale factor that is between 0.25 and 3.