Version Seven
For this version I added exceptional error control for the float and int inputs. I created two functions – force_int and force_num. Force_int will loop infinitely until the user enters a whole number, while force_num will do the same except the user can enter a whole number or a decimal. Either way, it’ll get converted to a float when it is stored. I replaced the input statements with these functions for the input of the ingredient amounts and the serving inputs (both old and new).