Version Eight
This version is what I plan to be the final version of this program. I’ve gone through and deleted any unnecessary comments and print statements to make the code easier to read. Usability wise, I added a welcome/instruction message so the user knows what to do. I tried to keep it simple to not overwhelm them, though, as I’ve tried to make the input prompts as descriptive as possible. Also, for the final recipe output, I made an adjustment so that when you write n/a for the unit it will format the output differently and not include any unit in that line. This allows for things like “2 eggs” and “3 cups of flour” to exist in the same recipe.